@@ -3,19 +3,24 @@ import QueryListing from "@/components/query-listing";
 import QueryPagination from "@/components/query-pagination";
 import SearchInput from "@/components/searchInput";
 import { getSearchParams } from "@/lib/utils";
+import { Metadata } from "next";
 
 type Props = {
   searchParams: searchParams;
 };
 
+export const metadata: Metadata = {
+  title: "These2.0 | Search",
+  description: "Search for title, author, category, tag, prof...",
+};
 const page = async ({ searchParams }: Props) => {
   const { categ, lang, search, sort, year, page } = await getSearchParams(
     searchParams
   );
-
   const isShowOptions = !!categ || !!lang || !!year;
 
   const fallBackData = { pageData: [], totalPages: 1, totalTheses: 0 };
+  // console.log("lang is: ",lang)
   const { pageData, totalPages, totalTheses } =
     (await getTheses({
       searchParams: { search: search, category: categ, langue: lang, year },
