@@ -18,17 +18,12 @@ const PdfFullScreen = ({ href, title }: Props) => {
   const [error, setError] = useState();
 
   const handleLoad = () => {
-    console.log("Embed is loaded");
-    console.log("handle loading: ", isPending);
     setIsPending(false); // Set isPending to false when loaded
   };
   const handleError = () => {
-    console.error("Error loading embed");
     setIsPending(false); // Set isPending to false on error
   };
-  useEffect(() => {
-    console.log(isPending);
-  }, [isPending]);
+
   return (
     <Dialog
       open={isOpen}
