@@ -1,7 +1,6 @@
-"use client"
+"use client";
 import { Toaster } from "@/components/ui/sonner";
 import React from "react";
-
 
 type Props = {
   children: React.ReactNode;
@@ -11,7 +10,7 @@ const Provider = ({ children }: Props) => {
   return (
     <>
       {children}
-      <Toaster  position="top-center"/>
+      <Toaster position="top-center" theme="light" />
     </>
   );
 };
